+++ src/playwright-azure-reporter.ts
@@ -5,13 +5,16 @@ import * as azdev from 'azure-devops-node-api';
 import { WebApi } from 'azure-devops-node-api';
 import { ICoreApi } from 'azure-devops-node-api/CoreApi';
 import { IRequestOptions } from 'azure-devops-node-api/interfaces/common/VsoBaseInterfaces';
+import VSSInterfaces from 'azure-devops-node-api/interfaces/common/VSSInterfaces';
 import { TeamProject } from 'azure-devops-node-api/interfaces/CoreInterfaces';
 import * as TestInterfaces from 'azure-devops-node-api/interfaces/TestInterfaces';
-import { TestPoint } from 'azure-devops-node-api/interfaces/TestInterfaces';
+import * as TestPlanInterfaces from 'azure-devops-node-api/interfaces/TestPlanInterfaces';
 import * as Test from 'azure-devops-node-api/TestApi';
+import * as TestPlanApi from 'azure-devops-node-api/TestPlanApi';
 import { setVariable } from 'azure-pipelines-task-lib';
 import chalk from 'chalk';
 import { existsSync, readFileSync } from 'fs';
+import * as restm from 'typed-rest-client/RestClient';
 import { isRegExp } from 'util/types';
 
 import Logger from './logger';
@@ -54,11 +57,16 @@ export interface AzureReporterOptions {
   uploadAttachments?: boolean | undefined;
   attachmentsType?: TAttachmentType | undefined;
   testRunConfig?: TTestRunConfig;
-  testPointMapper?: (testCase: TestCase, testPoints: TestPoint[]) => Promise<TestPoint[] | undefined>;
+  testPointMapper?: (
+    testCase: TestCase,
+    testPoints: TestInterfaces.TestPoint[]
+  ) => Promise<TestInterfaces.TestPoint[] | undefined>;
   isExistingTestRun?: boolean;
   testRunId?: number;
   testCaseIdMatcher?: string | RegExp | Array<string | RegExp>;
   testCaseIdZone?: TTestCaseIdZone;
+  rootSuiteId?: number;
+  uploadLogs?: boolean;
 }
 
 interface TestResultsToTestRun {
@@ -115,8 +123,15 @@ interface Headers {
 class AzureDevOpsReporter implements Reporter {
   private _logger: Logger | undefined;
   private _testApi!: Test.ITestApi;
+  private _testPlanApi!: TestPlanApi.ITestPlanApi;
   private _coreApi!: ICoreApi;
-  private _publishedResultsCount = 0;
+  private _publishedResultsCount: {
+    tests: number;
+    points: number;
+  } = {
+    tests: 0,
+    points: 0,
+  };
   private _testsAliasToBePublished: string[] = [];
   private _testResultsToBePublished: TTestResultsToBePublished[] = [];
   private _connection!: WebApi;
@@ -138,7 +153,10 @@ class AzureDevOpsReporter implements Reporter {
   private _resolvePublishResults: () => void = () => {};
   private _rejectPublishResults: (reason: any) => void = () => {};
   private _testRunConfig: TTestRunConfig = {} as TTestRunConfig;
-  private _testPointMapper: (testCase: TestCase, testPoints: TestPoint[]) => Promise<TestPoint[] | undefined>;
+  private _testPointMapper: (
+    testCase: TestCase,
+    testPoints: TestInterfaces.TestPoint[]
+  ) => Promise<TestInterfaces.TestPoint[] | undefined>;
   private _azureClientOptions = {
     allowRetries: true,
     maxRetries: 20,
@@ -148,6 +166,12 @@ class AzureDevOpsReporter implements Reporter {
   private _isExistingTestRun = false;
   private _testCaseIdMatcher: string | RegExp | Array<string | RegExp> = new RegExp(/\[([\d,\s]+)\]/, 'g');
   private _testCaseIdZone: TTestCaseIdZone = 'title';
+  private _rootSuiteId: number | undefined;
+  private _expectedTestPointsByRootSuite: Array<TestPlanInterfaces.TestPoint> = [];
+  private _testPointsByRootSuitePromise: Promise<boolean | void>;
+  private _resolveTestPointsByRootSuite: (value: boolean) => void = () => {};
+  private _rejectTestPointsByRootSuite: (reason: any) => void = () => {};
+  private _uploadLogs = false;
 
   public constructor(options: AzureReporterOptions) {
     this._runIdPromise = new Promise<number | void>((resolve, reject) => {
@@ -182,6 +206,21 @@ class AzureDevOpsReporter implements Reporter {
 
       return testPoints;
     };
+
+    this._rootSuiteId = options.rootSuiteId || Number(process.env.AZURE_PW_ROOT_SUITE_ID) || undefined;
+    this._testPointsByRootSuitePromise = new Promise<boolean | void>((resolve, reject) => {
+      this._resolveTestPointsByRootSuite = resolve;
+      this._rejectTestPointsByRootSuite = reject;
+    })
+      .then((state) => {
+        return state;
+      })
+      .catch((error) => {
+        this._logger?.error(error);
+        this._setIsDisable(true);
+      });
+    this._uploadLogs = options.uploadLogs || false;
+
     this._validateOptions(options);
   }
 
@@ -283,22 +322,31 @@ class AzureDevOpsReporter implements Reporter {
 
   async onBegin(): Promise<void> {
     if (this._isDisabled) return;
-    if (this._isExistingTestRun) {
-      this._resolveRunId(this._testRunId!);
-      this._logger?.log(`Using existing run ${this._testRunId} to publish test results`);
-      this._logger?.log(`AZURE_PW_TEST_RUN_ID: ${process.env.AZURE_PW_TEST_RUN_ID}`);
-      return;
-    }
     try {
+      if (this._isExistingTestRun) {
+        this._resolveRunId(this._testRunId!);
+        this._logger?.info(`Using existing run ${this._testRunId} to publish test results`);
+        this._logger?.info(`AZURE_PW_TEST_RUN_ID: ${process.env.AZURE_PW_TEST_RUN_ID}`);
+
+        if (this._rootSuiteId) {
+          await this._retrieveTestPoints();
+        }
+
+        return;
+      }
       this._testApi = await this._connection.getTestApi();
 
       if (this._publishTestResultsMode === 'testResult') {
         const run = await this._createRun(this._testRunTitle);
         if (run?.id) {
           this._resolveRunId(run.id);
-          this._logger?.log(`Using run ${run.id} to publish test results`);
+          this._logger?.info(`Using run ${run.id} to publish test results`);
           this._setAzurePWTestRunId(run.id);
-          this._logger?.log(`AZURE_PW_TEST_RUN_ID: ${process.env.AZURE_PW_TEST_RUN_ID}`);
+          this._logger?.info(`AZURE_PW_TEST_RUN_ID: ${process.env.AZURE_PW_TEST_RUN_ID}`);
+
+          if (this._rootSuiteId) {
+            await this._retrieveTestPoints();
+          }
         } else {
           this._rejectRunId('Failed to create test run. Reporting is disabled.');
         }
@@ -311,7 +359,7 @@ class AzureDevOpsReporter implements Reporter {
         this._logger?.error('Failed to create test run. Check your orgUrl. Reporting is disabled.');
       } else {
         this._logger?.error('Failed to create test run. Reporting is disabled.');
-        this._logger?.error(error.message);
+        this._logger?.error(error.stack);
       }
       this._setIsDisable(true);
     }
@@ -322,8 +370,11 @@ class AzureDevOpsReporter implements Reporter {
     try {
       if (this._publishTestResultsMode === 'testResult') {
         const runId = await this._runIdPromise;
-
         if (!runId) return;
+
+        if (this._rootSuiteId) {
+          await this._testPointsByRootSuitePromise;
+        }
 
         this._logTestItem(test, testResult);
         await this._publishCaseResult(test, testResult);
@@ -340,6 +391,7 @@ class AzureDevOpsReporter implements Reporter {
       }
     } catch (error: any) {
       this._logger?.error(`Failed to publish test result. \n ${error.message}`);
+      this._logger?.error(error.stack);
     }
   }
 
@@ -351,11 +403,15 @@ class AzureDevOpsReporter implements Reporter {
       if (this._publishTestResultsMode === 'testResult') {
         runId = await this._runIdPromise;
 
+        if (this._rootSuiteId) {
+          await this._testPointsByRootSuitePromise;
+        }
+
         let prevCount = this._testsAliasToBePublished.length;
         while (this._testsAliasToBePublished.length > 0) {
           // need wait all results to be published
           if (prevCount > this._testsAliasToBePublished.length) {
-            this._logger?.log(
+            this._logger?.info(
               `Waiting for all results to be published. Remaining ${this._testsAliasToBePublished.length} results`
             );
             prevCount--;
@@ -363,10 +419,8 @@ class AzureDevOpsReporter implements Reporter {
           await new Promise((resolve) => setTimeout(resolve, 250));
         }
       } else {
-        this._logger = new Logger(true);
-
         if (this._testResultsToBePublished.length === 0) {
-          this._logger?.log('No test results to publish');
+          this._logger?.info('No test results to publish', true);
           return;
         } else {
           if (!this._isExistingTestRun) {
@@ -377,10 +431,10 @@ class AzureDevOpsReporter implements Reporter {
           }
           if (runId) {
             this._resolveRunId(runId);
-            this._logger?.log(`Using run ${runId} to publish test results`);
+            this._logger?.info(`Using run ${runId} to publish test results`);
             this._setAzurePWTestRunId(runId);
-            this._logger?.log(`AZURE_PW_TEST_RUN_ID: ${process.env.AZURE_PW_TEST_RUN_ID}`);
-            await this._publishTestResults(runId, this._testResultsToBePublished);
+            this._logger?.info(`AZURE_PW_TEST_RUN_ID: ${process.env.AZURE_PW_TEST_RUN_ID}`);
+            await this._publishTestRunResults(runId, this._testResultsToBePublished);
           } else {
             this._setIsDisable(true);
             this._rejectRunId('Failed to create test run. Reporting is disabled.');
@@ -390,17 +444,22 @@ class AzureDevOpsReporter implements Reporter {
         }
       }
 
-      if (this._publishedResultsCount === 0 && !runId) {
-        this._logger?.log(chalk.gray('No testcases were matched. Ensure that your tests are declared correctly.'));
+      const shouldForceLogs = this._publishTestResultsMode === 'testRun';
+
+      if (this._publishedResultsCount.tests === 0 && !runId) {
+        this._logger?.warn('No testcases were matched. Ensure that your tests are declared correctly.');
         return;
+      } else {
+        const { tests, points } = this._publishedResultsCount;
+        this._logger?.info(`Test results published for ${tests} test(s), ${points} test point(s)`, shouldForceLogs);
       }
 
       if (this._isExistingTestRun) return;
       if (!this._testApi) this._testApi = await this._connection.getTestApi();
       const runUpdatedResponse = await this._testApi.updateTestRun({ state: 'Completed' }, this._projectName, runId!);
-      this._logger?.log(`Run ${runId} - ${runUpdatedResponse.state}`);
+      this._logger?.info(`Run ${runId} - ${runUpdatedResponse.state}`, shouldForceLogs);
     } catch (error: any) {
-      this._logger?.error(`Error on completing run ${error as string}`);
+      this._logger?.error(`Error on onEnd hook ${error as string}`);
     }
   }
 
@@ -519,22 +578,22 @@ class AzureDevOpsReporter implements Reporter {
   private _logTestItem(test: TestCase, testResult: TestResult) {
     switch (testResult.status) {
       case 'passed':
-        this._logger?.log(chalk.green(`${test.title} - ${testResult.status}`));
+        this._logger?.info(chalk.green(`${test.title} - ${testResult.status}`));
         break;
       case 'failed':
-        this._logger?.log(chalk.red(`${test.title} - ${testResult.status}`));
+        this._logger?.info(chalk.red(`${test.title} - ${testResult.status}`));
         break;
       case 'timedOut':
-        this._logger?.log(chalk.yellow(`${test.title} - ${testResult.status}`));
+        this._logger?.info(chalk.yellow(`${test.title} - ${testResult.status}`));
         break;
       case 'skipped':
-        this._logger?.log(chalk.yellow(`${test.title} - ${testResult.status}`));
+        this._logger?.info(chalk.yellow(`${test.title} - ${testResult.status}`));
         break;
       case 'interrupted':
-        this._logger?.log(chalk.red(`${test.title} - ${testResult.status}`));
+        this._logger?.info(chalk.red(`${test.title} - ${testResult.status}`));
         break;
       default:
-        this._logger?.log(`${test.title} - ${testResult.status}`);
+        this._logger?.info(`${test.title} - ${testResult.status}`);
         break;
     }
   }
@@ -561,7 +620,7 @@ class AzureDevOpsReporter implements Reporter {
         else throw new Error('Failed to create test run');
       }
     } catch (error: any) {
-      this._logger?.error(error.message);
+      this._logger?.error(error.stack);
       this._setIsDisable(true);
     }
   }
@@ -578,16 +637,140 @@ class AzureDevOpsReporter implements Reporter {
       if (project) return project;
       else throw new Error(`Project ${projectName} does not exist. Reporting is disabled.`);
     } catch (error: any) {
-      this._logger?.error(error.message);
+      this._logger?.error(error.stack);
       this._setIsDisable(true);
     }
   }
 
+  private _filterTestTestPoints(
+    testPoints: TestInterfaces.TestPoint[],
+    testsResult: TTestResultsToBePublished
+  ): TestInterfaces.TestPoint[] {
+    return testPoints?.filter((testPoint) => {
+      if (!testPoint.testPlan!.id || parseInt(testPoint.testPlan!.id, 10) !== this._planId) {
+        return false;
+      } else if (
+        !testPoint.testCase.id ||
+        !testsResult.testCase.testCaseIds.find((testCaseId) => testPoint.testCase.id == testCaseId)
+      ) {
+        return false;
+      } else if (this._testRunConfig && this._testRunConfig.configurationIds?.length) {
+        return (
+          testPoint.configuration.id &&
+          this._testRunConfig!.configurationIds.includes(parseInt(testPoint.configuration.id, 10))
+        );
+      } else {
+        return true;
+      }
+    });
+  }
+
+  private _filterTestPlanTestPoints(
+    testPoints: TestInterfaces.TestPoint[],
+    testsResult: TTestResultsToBePublished
+  ): TestInterfaces.TestPoint[] {
+    return testPoints?.filter((testPoint) => {
+      if (!testPoint.testPlan!.id || parseInt(testPoint.testPlan!.id, 10) !== this._planId) {
+        return false;
+      } else if (
+        !testPoint.testCase.id ||
+        !testsResult.testCase.testCaseIds.find((testCaseId) => testPoint.testCase.id == testCaseId)
+      ) {
+        return false;
+      } else if (this._testRunConfig && this._testRunConfig.configurationIds?.length) {
+        return (
+          testPoint.configuration.id &&
+          this._testRunConfig!.configurationIds.includes(parseInt(testPoint.configuration.id, 10))
+        );
+      } else {
+        return true;
+      }
+    });
+  }
+
+  private _convertTestPlanOutcomeToTestOutcome(outcome: TestPlanInterfaces.Outcome): string {
+    switch (outcome) {
+      case TestPlanInterfaces.Outcome.Passed:
+        return EAzureTestStatuses.passed;
+      case TestPlanInterfaces.Outcome.Failed:
+        return EAzureTestStatuses.failed;
+      case TestPlanInterfaces.Outcome.Inconclusive:
+        return EAzureTestStatuses.other;
+      case TestPlanInterfaces.Outcome.NotApplicable:
+        return EAzureTestStatuses.skipped;
+      case TestPlanInterfaces.Outcome.Paused:
+        return EAzureTestStatuses.fixme;
+      case TestPlanInterfaces.Outcome.Timeout:
+        return EAzureTestStatuses.timedOut;
+      case TestPlanInterfaces.Outcome.Aborted:
+        return EAzureTestStatuses.interrupted;
+      default:
+        return EAzureTestStatuses.other;
+    }
+  }
+
+  private _convertTestPlanTestPointsToTestPoints(
+    testPlanTestPoints: TestPlanInterfaces.TestPoint[]
+  ): TestInterfaces.TestPoint[] {
+    return testPlanTestPoints.map((testPlanTestPoint) => {
+      const {
+        isAutomated,
+        comment,
+        id,
+        project,
+        testCaseReference,
+        testPlan,
+        configuration,
+        lastUpdatedBy,
+        lastUpdatedDate,
+        links,
+        results,
+      } = testPlanTestPoint;
+      return {
+        assignedTo: testCaseReference.assignedTo,
+        automated: isAutomated,
+        comment: comment,
+        id: id,
+        projectId: project.id,
+        testCase: { id: testCaseReference.id.toString() },
+        testPlan: { id: testPlan.id.toString() },
+        configuration: { id: configuration.id.toString() },
+        lastUpdatedBy: lastUpdatedBy,
+        lastUpdatedDate: lastUpdatedDate,
+        outcome: this._convertTestPlanOutcomeToTestOutcome(results?.outcome),
+        url: links?._self.href || '',
+        workItemProperties: [],
+      } as TestInterfaces.TestPoint;
+    });
+  }
+
   private async _getTestPointsOfTestResults(
-    planId: number,
     testsResults: TTestResultsToBePublished[]
-  ): Promise<Map<TTestResultsToBePublished, TestPoint[]>> {
-    const result = new Map<TTestResultsToBePublished, TestPoint[]>();
+  ): Promise<Map<TTestResultsToBePublished, TestInterfaces.TestPoint[]>> {
+    const result = new Map<TTestResultsToBePublished, TestInterfaces.TestPoint[]>();
+
+    if (this._rootSuiteId) {
+      for (const testsResult of testsResults) {
+        const currentTestPoints = this._filterTestPlanTestPoints(
+          this._convertTestPlanTestPointsToTestPoints(this._expectedTestPointsByRootSuite),
+          testsResult
+        );
+
+        if (currentTestPoints && currentTestPoints.length > 0) {
+          const mappedTestPoints = await this._testPointMapper(testsResult.testCase, currentTestPoints);
+          if (mappedTestPoints && mappedTestPoints.length > 0) {
+            result.set(testsResult, mappedTestPoints);
+          } else {
+            result.set(testsResult, []);
+          }
+        } else {
+          result.set(testsResult, []);
+        }
+      }
+
+      return result;
+    }
+
     try {
       const testcaseIds = testsResults.map((t) => t.testCase.testCaseIds.map((id) => parseInt(id, 10))).flat();
       const pointsQuery: TestInterfaces.TestPointsQuery = {
@@ -600,27 +783,7 @@ class AzureDevOpsReporter implements Reporter {
       );
       if (pointsQueryResult.points) {
         for (const testsResult of testsResults) {
-          const currentTestPoints = pointsQueryResult.points?.filter((testPoint) => {
-            if (!testPoint.testPlan!.id || parseInt(testPoint.testPlan!.id, 10) !== planId) {
-              // the testPlan id is not matching
-              return false;
-            } else if (
-              !testPoint.testCase.id ||
-              !testsResult.testCase.testCaseIds.find((testCaseId) => testPoint.testCase.id == testCaseId)
-            ) {
-              // the testCase id is not matching
-              return false;
-            } else if (this._testRunConfig && this._testRunConfig.configurationIds?.length) {
-              // configuration ids are set, so they must match
-              return (
-                testPoint.configuration.id &&
-                this._testRunConfig!.configurationIds.includes(parseInt(testPoint.configuration.id, 10))
-              );
-            } else {
-              // no configuration ids to filter, ignore them
-              return true;
-            }
-          });
+          const currentTestPoints = this._filterTestTestPoints(pointsQueryResult.points, testsResult);
 
           if (currentTestPoints && currentTestPoints.length > 0) {
             const mappedTestPoints = await this._testPointMapper(testsResult.testCase, currentTestPoints);
@@ -628,17 +791,15 @@ class AzureDevOpsReporter implements Reporter {
             if (mappedTestPoints && mappedTestPoints.length > 0) {
               result.set(testsResult, mappedTestPoints);
             } else {
-              // logging will happen outside
               result.set(testsResult, []);
             }
           } else {
-            // logging will happen outside
             result.set(testsResult, []);
           }
         }
       }
     } catch (error: any) {
-      this._logger?.error(error.message);
+      this._logger?.error(error.stack);
     }
     return result;
   }
@@ -677,12 +838,96 @@ class AzureDevOpsReporter implements Reporter {
     return api;
   };
 
+  // prettier-ignore
+  private async _getPointsList(testPlanApi: TestPlanApi.ITestPlanApi, project: string, planId: number, suiteId: number, testPointIds?: string, testCaseId?: string, continuationToken?: string, returnIdentityRef?: boolean, includePointDetails?: boolean, isRecursive?: boolean): Promise<VSSInterfaces.PagedList<TestPlanInterfaces.TestPoint>> {
+    testPlanApi.getPointsList = function (project, planId, suiteId, testPointIds, testCaseId, continuationToken, returnIdentityRef, includePointDetails, isRecursive) {
+      return new Promise(async (resolve, reject) => {
+        try {
+          const routeValues = { project, planId, suiteId };
+          const queryValues = { testPointIds, testCaseId, continuationToken, returnIdentityRef, includePointDetails, isRecursive };
+          const verData = await testPlanApi.vsoClient.getVersioningData(
+            '7.2-preview.2',
+            'testplan',
+            '52df686e-bae4-4334-b0ee-b6cf4e6f6b73',
+            routeValues,
+            queryValues
+          );
+
+          let url: string = verData.requestUrl!;
+          let options = this.createRequestOptions('application/json', verData.apiVersion);
+
+          let res: restm.IRestResponse<VSSInterfaces.PagedList<TestPlanInterfaces.TestPoint>>;
+          res = await this.rest.get<VSSInterfaces.PagedList<TestPlanInterfaces.TestPoint>>(url, options);
+
+          let ret = this.formatResponse(res.result, TestPlanInterfaces.TypeInfo.TestPoint, true);
+
+          // Handle the continuation token
+          if (res.headers['x-ms-continuationtoken']) {
+            ret.continuationToken = res.headers['x-ms-continuationtoken'];
+          }
+
+          resolve(ret);
+        } catch (error: any) {
+          reject(error);
+        }
+      });
+    };
+
+    const testPlanApiWithOverride = testPlanApi as TestPlanApi.ITestPlanApi;
+    return testPlanApiWithOverride.getPointsList(project, planId, suiteId, testPointIds, testCaseId, continuationToken, returnIdentityRef, includePointDetails, isRecursive);
+  }
+
+  // prettier-ignore
+  private _recursivelyGetPointsList = async (project: string, planId: number, suiteId: number, continuationToken?: string): Promise<TestPlanInterfaces.TestPoint[]> => {
+    this._logger?.info(chalk.gray(`${continuationToken ? 'Fetching next' : 'Fetching'} test points.`));
+    try {
+      const testPlanApi = await this._connection.getTestPlanApi();
+      const pointsList = await this._getPointsList(testPlanApi, project, planId, suiteId, undefined, undefined, continuationToken, false, true, true);
+      let points: TestPlanInterfaces.TestPoint[] = pointsList;
+      if (pointsList?.continuationToken) {
+        const nextPoints = await this._recursivelyGetPointsList(project, planId, suiteId, pointsList?.continuationToken);
+        points = points.concat(nextPoints);
+      }
+      return points;
+    } catch (error: any) {
+      this._logger?.error(error.stack);
+      return [];
+    }
+  };
+
+  private async _retrieveTestPoints() {
+    if (this._rootSuiteId) {
+      this._logger?.info(`Fetching test points for the root suite ${this._rootSuiteId}`);
+      this._logger?.info('This may take a while...');
+      this._expectedTestPointsByRootSuite = await this._recursivelyGetPointsList(
+        this._projectName,
+        this._planId,
+        this._rootSuiteId
+      );
+      if (this._expectedTestPointsByRootSuite.length === 0) {
+        this._logger?.warn(
+          'No test points found for the specified root suite. Try to ensure that the rootSuiteId, planId are correct and the suite is not empty. Reporting is disabled.'
+        );
+        this._setIsDisable(true);
+      } else {
+        this._logger?.info(
+          `Fetched test points for the root suite ${this._rootSuiteId}. Total: ${this._expectedTestPointsByRootSuite.length}`
+        );
+        this._logger?.debug(
+          `First 3 Test points: ${JSON.stringify(this._expectedTestPointsByRootSuite.slice(0, 3), null, 2)}`
+        );
+      }
+
+      this._resolveTestPointsByRootSuite(true);
+    }
+  }
+
   private async _uploadAttachmentsFunc(
     testResult: TestResult,
     testCaseResultId: number,
     test: ITestCaseExtended | TestCase
   ): Promise<string[]> {
-    this._logger?.log(chalk.gray(`Uploading attachments for test: ${test.title}`));
+    this._logger?.info(chalk.gray(`Uploading attachments for test: ${test.title}`));
     const runId = await this._runIdPromise;
     const attachmentsResult: string[] = [];
 
@@ -723,10 +968,53 @@ class AzureDevOpsReporter implements Reporter {
           attachmentsResult.push(response.url);
         }
       } catch (error: any) {
-        this._logger?.error(error.message);
+        this._logger?.error(error.stack);
       }
     }
-    this._logger?.log(chalk.gray('Uploaded attachments'));
+    this._logger?.info(chalk.gray('Uploaded attachments'));
+    return attachmentsResult;
+  }
+
+  private async _uploadLogsAttachmentsFunc(
+    testResult: TestResult,
+    testCaseResultId: number,
+    test: ITestCaseExtended | TestCase
+  ): Promise<string[]> {
+    this._logger?.info(chalk.gray(`Uploading logs attachments for test: ${test.title}`));
+    const runId = await this._runIdPromise;
+    const attachmentsResult: string[] = [];
+
+    if (!runId) {
+      throw new Error('Could not find test run id. Check, maybe planId, what you specified, is incorrect.');
+    }
+
+    for (const attachment of testResult.attachments) {
+      try {
+        if (this._uploadLogs && ['stdout.txt', 'stderr.txt'].includes(attachment.name)) {
+          if (attachment.body) {
+            const attachmentRequestModel: TestInterfaces.TestAttachmentRequestModel = {
+              attachmentType: 'GeneralAttachment',
+              fileName: attachment.name,
+              stream: attachment.body?.toString('base64'),
+            };
+            if (!this._testApi) this._testApi = await this._connection.getTestApi();
+            const response = await this._testApi.createTestResultAttachment(
+              attachmentRequestModel,
+              this._projectName,
+              runId!,
+              testCaseResultId
+            );
+            if (!response?.id) throw new Error(`Failed to upload log attachment for test: ${test.title}`);
+            attachmentsResult.push(response.url);
+          } else {
+            throw new Error(`Attachment ${attachment.name} does not have body`);
+          }
+        }
+        this._logger?.info(chalk.gray('Uploaded logs attachments'));
+      } catch (error: any) {
+        this._logger?.error(error.stack);
+      }
+    }
     return attachmentsResult;
   }
 
@@ -746,6 +1034,53 @@ class AzureDevOpsReporter implements Reporter {
     }
   }
 
+  private _createTestCaseResult(
+    testCase: ITestCaseExtended,
+    testResult: TestResult,
+    testPoint: TestInterfaces.TestPoint | TestPlanInterfaces.TestPoint
+  ): TestInterfaces.TestCaseResult {
+    return {
+      testPoint: { id: String(testPoint.id) },
+      outcome: this._mapToAzureState(testCase, testResult),
+      state: 'Completed',
+      durationInMs: testResult.duration,
+      errorMessage: testResult.error
+        ? `${testCase.title}:\n\n${testResult.errors
+            ?.map((error, idx) => `ERROR #${idx + 1}:\n${error.message?.replace(/\u001b\[.*?m/g, '')}`)
+            .join('\n\n')}`
+        : undefined,
+      stackTrace: `${testResult.errors
+        ?.map((error, idx) => `STACK #${idx + 1}:\n\n${error.stack?.replace(/\u001b\[.*?m/g, '')}`)
+        .join('\n\n\n')}`,
+    };
+  }
+
+  private _prepareLogAttachments(testResult: TestResult) {
+    const logAttachments: {
+      name: string;
+      contentType: string;
+      body?: Buffer;
+    }[] = [];
+
+    if (testResult.stdout && testResult.stdout.length > 0) {
+      this._logger?.debug(`[prepareLogAttachments] stdout: ${testResult.stdout.join('')}`);
+      logAttachments.push({
+        name: 'stdout.txt',
+        contentType: 'text/plain',
+        body: Buffer.from(testResult.stdout.join('')),
+      });
+    }
+    if (testResult.stderr && testResult.stderr.length > 0) {
+      this._logger?.debug(`[prepareLogAttachments] stderr: ${testResult.stderr.join('')}`);
+      logAttachments.push({
+        name: 'stderr.txt',
+        contentType: 'text/plain',
+        body: Buffer.from(testResult.stderr.join('')),
+      });
+    }
+    return logAttachments;
+  }
+
   private async _publishCaseResult(test: TestCase, testResult: TestResult): Promise<TestResultsToTestRun | void> {
     const caseIds = this._getCaseIds(test);
     if (!caseIds || !caseIds.length) return;
@@ -760,11 +1095,10 @@ class AzureDevOpsReporter implements Reporter {
 
     try {
       const runId = await this._runIdPromise;
-      this._logger?.log(chalk.gray(`Start publishing: ${test.title}`));
+      this._logger?.info(chalk.gray(`Start publishing: ${test.title}`));
       const toBePublished: TTestResultsToBePublished = { testCase: testCase, testResult };
-      const mappedTestPoints = (await this._getTestPointsOfTestResults(this._planId as number, [toBePublished])).get(
-        toBePublished
-      );
+      const mappedTestPoints = (await this._getTestPointsOfTestResults([toBePublished])).get(toBePublished);
+      this._logger?.debug(`[publishCaseResult] Mapped test points: ${JSON.stringify(mappedTestPoints, null, 2)}`);
 
       if (!mappedTestPoints || mappedTestPoints.length == 0) {
         throw new Error(
@@ -772,27 +1106,13 @@ class AzureDevOpsReporter implements Reporter {
             this._testRunConfig?.configurationIds
               ? `for configurations [${this._testRunConfig?.configurationIds?.join(', ')}]`
               : ''
-          }\nCheck, maybe testPlanId or assigned configurations per test case, what you specified, is incorrect.`
+          }\nCheck, maybe testPlanId or assigned configurations per test case, what you specified, is incorrect.\n Also check testPointMapper function.`
         );
       }
+      this._publishedResultsCount.points += mappedTestPoints?.length || 0;
 
-      const results: TestInterfaces.TestCaseResult[] = mappedTestPoints.map(
-        (testPoint) =>
-          ({
-            // the testPoint is the testCase + configuration, there is not need to set these
-            testPoint: { id: String(testPoint.id) },
-            outcome: this._mapToAzureState(test, testResult),
-            state: 'Completed',
-            durationInMs: testResult.duration,
-            errorMessage: testResult.error
-              ? `${test.title}:\n\n${testResult.errors
-                  ?.map((error, idx) => `ERROR #${idx + 1}:\n${error.message?.replace(/\u001b\[.*?m/g, '')}`)
-                  .join('\n\n')}`
-              : undefined,
-            stackTrace: `${testResult.errors
-              ?.map((error, idx) => `STACK #${idx + 1}:\n\n${error.stack?.replace(/\u001b\[.*?m/g, '')}`)
-              .join('\n\n\n')}`,
-          } as TestInterfaces.TestCaseResult)
+      const results: TestInterfaces.TestCaseResult[] = mappedTestPoints.map((testPoint) =>
+        this._createTestCaseResult(testCase, testResult, testPoint)
       );
 
       if (!this._testApi) this._testApi = await this._connection.getTestApi();
@@ -805,17 +1125,27 @@ class AzureDevOpsReporter implements Reporter {
       if (this._uploadAttachments && testResult.attachments.length > 0)
         await this._uploadAttachmentsFunc(testResult, testCaseResult.result.value![0].id, test);
 
+      if (this._uploadLogs) {
+        this._logger?.info(chalk.gray(`Uploading stdout.log for test: ${test.title}`));
+        const logAttachment = this._prepareLogAttachments(testResult);
+        await this._uploadLogsAttachmentsFunc(
+          { ...testResult, attachments: logAttachment },
+          testCaseResult.result.value![0].id,
+          test
+        );
+      }
+
       this._removePublished(testCase.testAlias);
-      this._publishedResultsCount++;
-      this._logger?.log(chalk.gray(`Result published: ${test.title}`));
+      this._publishedResultsCount.tests++;
+      this._logger?.info(chalk.gray(`Result published: ${test.title}`));
       return testCaseResult;
     } catch (error: any) {
       this._removePublished(testCase.testAlias);
-      this._logger?.error(error.message);
+      this._logger?.error(error.stack);
     }
   }
 
-  private async _publishTestResults(runId: number, testsResults: TTestResultsToBePublished[]) {
+  private async _publishTestRunResults(runId: number, testsResults: TTestResultsToBePublished[]) {
     if (!this._testApi) this._testApi = await this._connection.getTestApi();
 
     const testsPackSize = 50;
@@ -824,13 +1154,24 @@ class AzureDevOpsReporter implements Reporter {
       testsResults.slice(i * testsPackSize, (i + 1) * testsPackSize)
     );
 
-    this._logger?.log(chalk.gray(`Start publishing test results for ${testsResults.length} test(s)`));
+    let testResultsQuery: TestInterfaces.TestResultsQuery;
+    let resultData: TestInterfaces.TestResultsQuery = { results: [] };
+
+    this._logger?.info(chalk.gray(`Start publishing test results for ${testsResults.length} test(s)`), true);
+
+    if (this._rootSuiteId) {
+      await this._retrieveTestPoints();
+      await this._testPointsByRootSuitePromise;
+    }
 
     try {
       for (const testsPack of testsPacksArray) {
         let testCaseIds: string[] = [];
-        const withAttachmentsByTestPoint = new Map<TestPoint, TTestResultsToBePublished[]>();
-        const testsPointsByTestCase = await this._getTestPointsOfTestResults(this._planId as number, testsPack);
+        const withAttachmentsByTestPoint = new Map<
+          TestInterfaces.TestPoint | TestPlanInterfaces.TestPoint,
+          TTestResultsToBePublished[]
+        >();
+        const testsPointsByTestCase = await this._getTestPointsOfTestResults(testsPack);
         const testCaseResults: TestInterfaces.TestCaseResult[] = [];
 
         for (const [key, value] of testsPointsByTestCase.entries()) {
@@ -841,21 +1182,7 @@ class AzureDevOpsReporter implements Reporter {
           if (testPoints && testPoints.length > 0) {
             testCaseIds.push(...testCase.testCaseIds);
             testCaseResults.push(
-              ...testPoints.map((testPoint) => ({
-                // the testPoint is the testCase + configuration, there is not need to set these
-                testPoint: { id: String(testPoint.id) },
-                outcome: this._mapToAzureState(testCase, testResult),
-                state: 'Completed',
-                durationInMs: testResult.duration,
-                errorMessage: testResult.error
-                  ? `${testCase.title}:\n\n${testResult.errors
-                      ?.map((error, idx) => `ERROR #${idx + 1}:\n${error.message?.replace(/\u001b\[.*?m/g, '')}`)
-                      .join('\n\n')}`
-                  : undefined,
-                stackTrace: `${testResult.errors
-                  ?.map((error, idx) => `STACK #${idx + 1}:\n\n${error.stack?.replace(/\u001b\[.*?m/g, '')}`)
-                  .join('\n\n\n')}`,
-              }))
+              ...testPoints.map((testPoint) => this._createTestCaseResult(testCase, testResult, testPoint))
             );
 
             if (this._uploadAttachments && testResult.attachments.length > 0) {
@@ -876,7 +1203,7 @@ class AzureDevOpsReporter implements Reporter {
                 this._testRunConfig?.configurationIds
                   ? `for configurations [${this._testRunConfig?.configurationIds?.join(', ')}]`
                   : ''
-              }\nCheck, maybe testPlanId or assigned configurations per test case, what you specified, is incorrect.`
+              }\nCheck, maybe testPlanId or assigned configurations per test case, what you specified, is incorrect.\n Also check testPointMapper function.`
             );
           }
         }
@@ -884,6 +1211,7 @@ class AzureDevOpsReporter implements Reporter {
         if (testCaseResults.length === 0) {
           continue;
         }
+        this._publishedResultsCount.points += testCaseResults.length;
 
         const testCaseResult: TestResultsToTestRun = (await this._addReportingOverride(
           this._testApi
@@ -895,19 +1223,20 @@ class AzureDevOpsReporter implements Reporter {
           this._logger?.warn(`Not all test result for test cases [${testCaseIds.join(', ')}] are published`);
         }
 
-        if (this._uploadAttachments && withAttachmentsByTestPoint.size > 0) {
-          this._logger?.log(
-            chalk.gray(`Starting to uploading attachments for ${withAttachmentsByTestPoint.size} testpoint(s)`)
-          );
-
-          const testResultsQuery: TestInterfaces.TestResultsQuery = {
+        if ((this._uploadAttachments && withAttachmentsByTestPoint.size > 0) || this._uploadLogs) {
+          testResultsQuery = {
             fields: [''],
             results: testCaseResult.result.value?.map((r) => {
               return { id: r.id, testRun: { id: r.testRun?.id } };
             }),
           };
+          resultData = await this._testApi.getTestResultsByQuery(testResultsQuery, this._projectName);
+        }
 
-          const resultData = await this._testApi.getTestResultsByQuery(testResultsQuery, this._projectName);
+        if (this._uploadAttachments && withAttachmentsByTestPoint.size > 0) {
+          this._logger?.info(
+            chalk.gray(`Starting to uploading attachments for ${withAttachmentsByTestPoint.size} testpoint(s)`)
+          );
 
           for (const [key, value] of withAttachmentsByTestPoint.entries()) {
             const testResult = resultData.results?.find((result) => result.testPoint?.id === String(key.id));
@@ -923,13 +1252,33 @@ class AzureDevOpsReporter implements Reporter {
           }
         }
 
-        this._publishedResultsCount += testsPack.length;
-        this._logger?.log(chalk.gray(`Left to publish: ${testsResults.length - this._publishedResultsCount}`));
+        if (this._uploadLogs && testsPack.length > 0) {
+          this._logger?.info(chalk.gray(`Starting to uploading logs for ${testsPack.length} test(s)`));
+          for (const [testWithResult, testPoints] of testsPointsByTestCase.entries()) {
+            for (const testPoint of testPoints) {
+              const testResult = resultData.results?.find((result) => result.testPoint?.id === String(testPoint.id));
+              if (!testResult) {
+                this._logger?.warn(`Test result for test point [${testPoint.id}] is missing, logs are not uploaded!`);
+                continue;
+              }
+
+              const logAttachment = this._prepareLogAttachments(testWithResult.testResult);
+              await this._uploadLogsAttachmentsFunc(
+                { ...testWithResult.testResult, attachments: logAttachment },
+                testResult.id!,
+                testWithResult.testCase
+              );
+            }
+          }
+        }
+
+        this._publishedResultsCount.tests += testsPack.length;
+        this._logger?.info(chalk.gray(`Left to publish: ${testsResults.length - this._publishedResultsCount.tests}`));
       }
-      this._logger?.log(chalk.gray(`Test results published for ${this._publishedResultsCount} test(s)`));
+
       this._resolvePublishResults();
     } catch (error: any) {
-      this._logger?.error(error.message);
+      this._logger?.error(error.stack);
       this._rejectPublishResults(error);
     }
   }